refactor: 헥사고날 아키텍처 -> 레이어드 아키텍처
#16
