+++ src/main/java/com/moa/global/BaseEnum.java
@@ -1,0 +1,9 @@
+package com.moa.global;
+
+
+public interface BaseEnum<T, K> {
+
+	T getCode();
+	K getTitle();
+
+}
