--- src/main/java/com/moa/meeting/adaptor/web/response/MeetingCreateRes.java
@@ -1,12 +1,0 @@
-package com.moa.meeting.adaptor.web.response;
-
-
-import lombok.Getter;
-
-
-@Getter
-public class MeetingCreateRes {
-
-	private Long id;
-
-}
