+++ src/main/java/com/moa/meeting/application/MeetingService.java
@@ -1,0 +1,12 @@
+package com.moa.meeting.application;
+
+
+import com.moa.meeting.dto.MeetingCreateDto;
+import com.moa.meeting.dto.MeetingGetDto;
+
+
+public interface MeetingService {
+
+	MeetingGetDto createMeeting(MeetingCreateDto meetingCreateDto);
+
+}
