--- src/main/java/com/moa/meeting/common/BaseEnum.java
@@ -1,9 +1,0 @@
-package com.moa.meeting.common;
-
-
-public interface BaseEnum<T, K> {
-
-	T getCode();
-	K getTitle();
-
-}
