--- src/main/java/com/moa/meeting/common/exception/CustomException.java
@@ -1,14 +1,0 @@
-package com.moa.meeting.common.exception;
-
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-
-@Getter
-@RequiredArgsConstructor
-public class CustomException extends RuntimeException {
-
-	private final ErrorCode errorCode;
-
-}
