+++ src/main/java/com/moa/meeting/infrastructure/mysql/MeetingRepository.java
@@ -1,0 +1,9 @@
+package com.moa.meeting.infrastructure.mysql;
+
+
+import com.moa.meeting.domain.Meeting;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+
+public interface MeetingRepository extends JpaRepository<Meeting, Long> {
+}
