+++ src/main/java/com/moa/meeting/presentation/MeetingController.java
@@ -1,0 +1,53 @@
+package com.moa.meeting.presentation;
+
+
+import com.moa.global.vo.ApiResult;
+import com.moa.meeting.application.MeetingService;
+import com.moa.meeting.dto.MeetingCreateDto;
+import com.moa.meeting.dto.MeetingGetDto;
+import com.moa.meeting.vo.request.MeetingCreateRequest;
+import com.moa.meeting.vo.response.MeetingCreateResponse;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.media.Schema;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.modelmapper.ModelMapper;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.util.UUID;
+
+
+@RestController
+@RequestMapping("/api/v1/meeting")
+@Slf4j
+@RequiredArgsConstructor
+public class MeetingController {
+
+	private final ModelMapper modelMapper;
+	private final MeetingService meetingService;
+
+
+	@Operation(summary = "모임 생성", description = "모임 생성")
+	@ApiResponses({
+		@ApiResponse(responseCode = "200", description = "OK",
+			content = @Content(schema = @Schema(implementation = MeetingCreateResponse.class))),
+		@ApiResponse(responseCode = "500", description = "INTERNAL SERVER ERROR")
+	})
+	@PostMapping("")
+	public ResponseEntity<ApiResult<MeetingCreateResponse>> createMeeting(@RequestBody MeetingCreateRequest request) {
+		log.debug(request.toString());
+		UUID userUuid = UUID.fromString("a642406c-6e20-11ee-b962-0242ac120002"); // TODO: 로그인 구현 후 수정
+		MeetingGetDto meetingGetDto = meetingService.createMeeting(modelMapper.map(request, MeetingCreateDto.class));
+		log.debug(meetingGetDto.toString());
+		MeetingCreateResponse response = modelMapper.map(meetingGetDto, MeetingCreateResponse.class);
+		return ResponseEntity.ok(ApiResult.ofSuccess(response));
+	}
+
+}
