+++ src/main/java/com/moa/meeting/vo/response/MeetingCreateResponse.java
@@ -1,0 +1,12 @@
+package com.moa.meeting.vo.response;
+
+
+import lombok.Getter;
+
+
+@Getter
+public class MeetingCreateResponse {
+
+	private Long id;
+
+}
